+++ src/components/AccountInfoGauge/DepositWithdrawModal/styles.less
@@ -1,0 +1,18 @@
+.ConfirmationBox {
+    font-family: "Karla", sans-serif;
+    font-size: 20px;
+}
+
+.ConfirmationBoxButton {
+    margin-top: 40px;
+    margin-bottom: 20px;
+    width: 100%;
+    height: 40px;
+    color: #b5b5b6;
+    background: transparent;
+    border: 0.75px solid #232323;
+    border-radius: 3px;
+    text-align: center;
+    font-size: 16px;
+    cursor: pointer;
+}
